separate out catch and photo quests, cleanup
